# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Resell API",
    "version": "1.0.0",
    "description": "API for the resell application"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/user/create": {
      "post": {
        "tags": ["User"],
        "summary": "Create a new user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      }
    },
    "/user": {
      "get": {
        "tags": ["User"],
        "summary": "Get all users",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/User" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["User"],
        "summary": "Edit user profile",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EditProfileRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["User"],
        "summary": "Delete current user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User deleted successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      }
    },
    "/user/id/{id}": {
      "get": {
        "tags": ["User"],
        "summary": "Get user by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Firebase UID of the user to delete",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["User"],
        "summary": "Delete user by other user",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Firebase UID of the user to delete",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/googleId/{id}": {
      "get": {
        "tags": ["User"],
        "summary": "Get user by Google ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/postId/{id}": {
      "get": {
        "tags": ["User"],
        "summary": "Get user by post ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/email": {
      "post": {
        "tags": ["User"],
        "summary": "Get user by email",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GetUserByEmailRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/admin": {
      "post": {
        "tags": ["User"],
        "summary": "Set user as admin",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAdminByEmailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User admin status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/block": {
      "post": {
        "tags": ["User"],
        "summary": "Block a user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BlockUserRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User blocked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/unblock": {
      "post": {
        "tags": ["User"],
        "summary": "Unblock a user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UnblockUserRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User unblocked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/blocked/id/{id}": {
      "get": {
        "tags": ["User"],
        "summary": "Get blocked users by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Firebase UID of the user",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Blocked users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/User" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/softDelete/id/{id}": {
      "post": {
        "tags": ["User"],
        "summary": "Soft delete user",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Firebase UID of the user to soft delete",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "User soft deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/logout": {
      "post": {
        "tags": ["User"],
        "summary": "Logout user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FcmTokenRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged out successfully"
          }
        }
      }
    },
    "/notif": {
      "post": {
        "summary": "Send notification",
        "tags": ["Notification"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FindTokensRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification sent successfully"
          }
        }
      }
    },
    "/notif/recent": {
      "get": {
        "tags": ["Notification"],
        "summary": "Get recent notifications",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Recent notifications retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Notification" }
                }
              }
            }
          }
        }
      }
    },
    "/notif/request-match": {
      "post": {
        "summary": "Send request match notification",
        "tags": ["Notification"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestMatchNotificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request match notification sent successfully"
          }
        }
      }
    },
    "/notif/discount": {
      "post": {
        "summary": "Send discount notification",
        "tags": ["Notification"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscountNotificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Discount notification sent successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateUserRequest": {
        "type": "object",
        "required": [
          "username",
          "netid",
          "givenName",
          "familyName",
          "photoUrl",
          "venmoHandle",
          "email",
          "googleId",
          "bio",
          "fcmToken"
        ],
        "properties": {
          "username": { "type": "string" },
          "netid": { "type": "string" },
          "givenName": { "type": "string" },
          "familyName": { "type": "string" },
          "photoUrl": { "type": "string" },
          "venmoHandle": { "type": "string" },
          "email": { "type": "string" },
          "googleId": { "type": "string" },
          "bio": { "type": "string" },
          "fcmToken": { "type": "string" }
        }
      },
      "EditProfileRequest": {
        "type": "object",
        "properties": {
          "photoUrlBase64": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "venmoHandle": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "GetUserByEmailRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string" }
        },
        "required": ["email"]
      },
...
      "FcmTokenRequest": {
        "type": "object",
        "properties": {
          "token": { "type": "string" }
        },
        "required": ["token"]
      },
      "User": {
        "type": "object",
        "properties": {
          "firebaseUid": { "type": "string" },
          "username": { "type": "string" },
          "netid": { "type": "string" },
          "givenName": { "type": "string" },
          "familyName": { "type": "string" },
          "stars": { "type": "number" },
          "numReviews": { "type": "number" },
          "photoUrl": { "type": "string" },
          "venmoHandle": { "type": "string" },
          "bio": { "type": "string" },
          "posts": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Post" }
          },
          "admin": { "type": "boolean" },
          "email": { "type": "string" },
          "googleId": { "type": "string" },
          "isActive": { "type": "boolean" },
          "feedbacks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Feedback" }
          },
          "blockers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/User" }
          },
          "blocking": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/User" }
          },
          "reports": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Report" }
          },
          "reportedBy": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Report" }
          }
        }
      },
      "FindTokensRequest": {
        "type": "object",
        "required": ["email", "title", "body", "data"],
        "properties": {
          "email": { "type": "string" },
          "title": { "type": "string" },
          "body": { "type": "string" },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "RequestMatchNotificationRequest": {
        "type": "object",
        "required": ["requestId", "listingId", "userId"],
        "properties": {
          "requestId": { "type": "string" },
          "listingId": { "type": "string" },
          "userId": { "type": "string" }
        }
      },
      "DiscountNotificationRequest": {
        "type": "object",
        "properties": {
          "sellerId": {
            "type": "string",
            "description": "Firebase UID of the seller"
          },
          "listingId": {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the post/listing"
          },
          "oldPrice": { "type": "number" },
          "newPrice": { "type": "number" }
        },
        "required": ["sellerId", "listingId", "oldPrice", "newPrice"]
      },
      "Notification": {
        "type": "object",
        "required": ["to", "sound", "title", "body", "data"],
        "properties": {
          "to": {
            "type": "array",
            "items": { "type": "string" }
          },
          "sound": {
            "type": "string",
            "enum": ["default"]
          },
          "title": { "type": "string" },
          "body": { "type": "string" },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "Report": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "reporter": { "$ref": "#/components/schemas/PrivateProfile" },
          "reported": { "$ref": "#/components/schemas/PrivateProfile" },
          "post": {
            "type": "object",
            "nullable": true,
            "$ref": "#/components/schemas/Post"
          },
          "message": {
            "type": "object",
            "nullable": true,
            "$ref": "#/components/schemas/Message"
          },
          "reason": { "type": "string" },
          "type": {
            "type": "string",
            "enum": ["post", "profile", "message"]
          },
          "resolved": { "type": "boolean" },
          "created": { "type": "string", "format": "date-time" }
        }
      },
...
      "SetAdminByEmailRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "status": { "type": "boolean" }
        },
        "required": ["email", "status"]
      },
      "BlockUserRequest": {
        "type": "object",
        "properties": {
          "blocked": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["blocked"]
      },
      "UnblockUserRequest": {
        "type": "object",
        "properties": {
          "unblocked": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["unblocked"]
      },
...
      "PrivateProfile": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "firebaseUid": { "type": "string" },
          "username": { "type": "string" },
          "netid": { "type": "string" },
          "givenName": { "type": "string" },
          "familyName": { "type": "string" },
          "photoUrl": { "type": "string" },
          "venmoHandle": { "type": "string" },
          "email": { "type": "string" },
          "googleId": { "type": "string" },
          "bio": { "type": "string" },
          "admin": { "type": "boolean" },
          "created": { "type": "string", "format": "date-time" },
          "updated": { "type": "string", "format": "date-time" }
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "category": { "type": "string" },
          "condition": { "type": "string" },
          "originalPrice": { "type": "number" },
          "alteredPrice": { "type": "number" },
          "images": {
            "type": "array",
            "items": { "type": "string" }
          },
          "created": { "type": "string", "format": "date-time" },
          "location": { "type": "string" },
          "archive": { "type": "boolean" },
          "user": { "$ref": "#/components/schemas/User" },
          "savers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PrivateProfile" }
          },
          "matched": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Request" }
          },
          "categories": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Category" }
          },
          "sold": { "type": "boolean" },
          "reports": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Report" }
          }
        }
      },
      "Feedback": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "description": { "type": "string" },
          "images": {
            "type": "array",
            "items": { "type": "string" }
          },
          "created": { "type": "string", "format": "date-time" },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "content": { "type": "string" },
          "sender": { "$ref": "#/components/schemas/PrivateProfile" },
          "receiver": { "$ref": "#/components/schemas/PrivateProfile" },
          "created": { "type": "string", "format": "date-time" },
          "read": { "type": "boolean" }
        }
      },
      "Request": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "status": { "type": "string" },
          "created": { "type": "string", "format": "date-time" },
          "updated": { "type": "string", "format": "date-time" },
          "buyer": { "$ref": "#/components/schemas/PrivateProfile" },
          "post": { "$ref": "#/components/schemas/Post" }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "posts": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Post" }
          }
        }
      },
...
    }
  }
}